Added Box components that are moveable within the AppSpace
