+++ src/components/AppSpace/AppSpace.css
@@ -1,4 +1,4 @@
-.appSpace-container {
+.App {
     border: 4px solid green;
     height: 100vh;
     width: 95%;
