+++ src/components/Box/Box.jsx
@@ -1,0 +1,66 @@
+import React, { useState, useRef, useEffect } from 'react';
+
+function Box({ initialPosition }) {
+  const [position, setPosition] = useState(initialPosition);
+  const [isDragging, setIsDragging] = useState(false);
+  const boxRef = useRef(null);
+
+  useEffect(() => {
+    const handleMouseUp = () => {
+      if (isDragging) {
+        setIsDragging(false);
+      }
+    };
+
+    window.addEventListener('mouseup', handleMouseUp);
+
+    return () => {
+      window.removeEventListener('mouseup', handleMouseUp);
+    };
+  }, [isDragging]);
+
+  const handleMouseDown = (event) => {
+    if (boxRef.current.contains(event.target)) {
+      setIsDragging(true);
+    }
+  };
+
+  const handleMouseMove = (event) => {
+    if (isDragging) {
+      const appRect = document.body.getBoundingClientRect();
+      const boxRect = boxRef.current.getBoundingClientRect();
+      const maxX = appRect.width - boxRect.width;
+      const maxY = appRect.height - boxRect.height;
+      let newX = position.x + event.movementX;
+      let newY = position.y + event.movementY;
+      newX = Math.min(Math.max(newX, 0), maxX);
+      newY = Math.min(Math.max(newY, 0), maxY);
+      const newPosition = { x: newX, y: newY };
+      setPosition(newPosition);
+    }
+  };
+
+  const boxStyle = {
+    position: 'absolute',
+    left: position.x + 'px',
+    top: position.y + 'px',
+    width: '100px',
+    height: '100px',
+    backgroundColor: 'red',
+    cursor: isDragging ? 'move' : 'default',
+  };
+
+  return (
+    <div
+      className="box"
+      ref={boxRef}
+      style={boxStyle}
+      onMouseDown={handleMouseDown}
+      onMouseMove={handleMouseMove}
+    >
+      component
+    </div>
+  );
+}
+
+export default Box;
